\name{dataelev}
\alias{dataelev}
\docType{data}
\title{Surface elevations}
\description{
Surface elevation data taken from Davis (1972). An onject of the class geodata with elevation values at 52 locations.
}
\usage{data("dataelev")}
\format{
  A data frame with 52 observations on the following 3 variables.
  \describe{
    \item{\code{x}}{X coordinate (multiple of 50 feet).}
    \item{\code{y}}{Y coordinate (multiple of 50 feet).}
    \item{\code{elevation}}{elevations (multiples of 10 feet).}
  }
}

\references{
Davis, J.C. (1973) Statistics and Data Analysis in Geology. Wiley.
}

